BackendAssignment/requirements.txt
requests==2.14.2
python-firebase==1.2
